IDPROD-2749: Error Screen (#460)
Added a basic `ErrorViewController` to our the native Identity Verification flow. This screen still needs to be updated to match the design (once it's been finalized) but displays the basics. 

Currently, the VC will pop the navigation stack when the user hits the button, although we may need to add some more sophisticated logic in the future for certain errors (like camera permissions). The idea is that it will show backend validation errors like, the user entered a Date of Birth that shows they're underage, so we'll display an error message and give them the opportunity to go back to the previous screen and re-enter the data.

Also added some mocking capabilities to force the error state to happen on a specific screen. It's not very pretty since I had to add a stepper in the example app to say which screen the error should be injected into, but it's only temporarily needed because our endpoint is not live yet.