IDPROD-2740: Document selection persistence & screen routing (#464)
Added a Document selection screen to the IDV flow. The purpose of this screen is for the user to select what type of identification they're going to scan prior to scanning it (e.g. passport, driver's license, id card).

Note: The UI does not match the latest design but rather was the quickest UI to enable this view in the flow. Since the design is still in flux, we will address updating the UI to match design in a later milestone.

We've future-proofed the API to be flexible to potential future ID types that the client doesn't currently know about beyond ID Cards, passports, and drivers licenses by sending a string-string dictionary of ID-type + localized description. However, for our initial client release, we're only going to be able to scan these 3 identification types. So the `DocumentTypeSelectViewController` has some logic to filter down the selection sent by the API to only known types and default to these 3 options if no valid values are sent from the server.

#### Testing & Mocks

I also beefed up our testing infrastructure by making the `VerificationSheetFlowController` and `VerificationSheetController` dependency-injectable. There are now protocols and mock versions of these classes to aid in functional testing. See `DocumentTypeSelectViewControllerTest` for how this is used.

Side-note: I'll probably go back and add a `BiometricConsentViewControllerTest` in a future PR now that we have these test mock classes.

#### File organization

Also did some minor file organization for IdentityTests so each test mirrors the same file hierarchy structure as the class its testing uses in the Source folder.