IDPROD-2742 part 2: ID Scan ViewController (#481)
Added a `DocumentCaptureViewController` to scaffold out scanning an ID document. If the user selected that they wanted to scan a driver's license or ID card on the previous screen, we'll prompt to scan the front & back of the card. If they selected passport then we only scan the front of the document.

#### Mocks

The work to get configure the camera and scan the camera feed for specific documents is slated for another milestone. So I've created a simple `DocumentScanner` to scan a `CVPixelBuffer` for a valid document. Currently, it just passes back the original pixelBuffer after 5 seconds. 

I've also created a `MockIdentityDocumentCameraFeed` class to send a `CVPixelBuffer` from an image loaded from the file system. This might change a bit when we build out camera feed logic, but the idea is that we'll be able to reuse it to mock the camera feed in tests as well as when testing on the simulator. Right now, it returns a front ID image followed by a back ID image that's mocked out in the example app.