IDPROD-2742  part 1: Migrate uploadImage to StripeCore (#479)
Migrated logic for `STPAPIClient.uploadImage` to `StripeCore` so `StripeIdentity` could upload images of identity documents.

The `@objc` public method, `uploadImage(_ image: UIImage, purpose: STPFilePurpose, completion: STPFileCompletionBlock?)` still remains in the `Stripe` module with no signature change. This is because `STPFileCompletionBlock` has a dependency on `STPAPIResponseDecodable`, so it must remain in `Stripe`.

I created a new `StripeFile` struct in `StripeCore` which uses our new `StripeDecodable` swifty-decoding. There's now an SPI-public version of `uploadImage()` which uses `StripeFile`. The public version of `uploadImage` in `Stripe` has been refactored to wrap the SPI-public method version and transform the completion result from `StripeFile` -> `STPFile`.

The business logic for `uploadImage()` was migrated to the SPI-public `StripeCore` method, with some minor modifications:
- It's been simplified to force-unwrap the URL instantiation. The old biz logic would instead return without executing the completion block. (We force-unwrap the URLs for most of our API calls that use const URL strings.)
- It DRYs up the logic to use `sendRequest` instead of repeating it inside `stp_performDataTask`
- Using `HTTPMethod.post.rawValue` instead of `"POST"` for `httpMethod`