IDPROD-2742 part 4: Upload ID Document Images (#492)
Now Identity document images are uploaded to our file upload endpoint as soon as the images are captured. When the user hits "continue" to finish the final step in document capture, the ViewController will:
1. Wait until the images are finished uploading (if they haven't finished yet)
2. Update the data store with the images and the resulting file ID
3. Save the IDs of the uploaded files to the `VerificationSessionData` endpoint

#### Saving Images
When testing on the simulator, I observed that it takes ~7 seconds for the files to finish uploading. We likely will need to downscale the images prior to uploading so it doesn't take so long (part of IDPROD-2482).

We begin to upload the image as soon as we get a valid image capture from the camera feed, before the user confirms the image. In the event that the images are not done uploading by the time the user hits the final "Continue" button to go to the next screen, I added a `.saving` state to the view controller that will disable the button until it's done uploading and saving the data. We'll need to add a better loading state when we cleanup the UI to match the design (IDPROD-2756).

#### Misc

- Also added a `StripeFileMock` to `StripeCoreTestingUtils` to mock a file upload response.
- Updated `VerificationSheetControllerTest` to load JSON mocks during setup as opposed to at the top of the file, since this effects our test env