IDPROD-2770: VerificationSession submit endpoint (#512)
Added ability to post to the `v1/verification_session/:id/submit` endpoint. When the user has finished entering all of the required fields as part of the verification flow, we post to this endpoint to signal to the API that the user has finished submitting all of their data. If the response returns without errors, then we display a success state that tells the user their verification is pending.