IDPROD-2855: Success screen (#522)
Display a success screen at the end of the verification flow, after the user's data has been submitted.

Did some minor refactoring to `VerificationSheetFlowController.transitionToNextScreen`. Consolidated the logic between 
`transitionToFirstScreen` and `transitionToNextScreen` then used the state of the navigation stack + next VC to determine whether to replace the navigation stack vs. push. The design shows that the user should not be able to go back after they've already submitted their data, so we want to replace the VC stack when they're done.