IDPROD-2953 part 2: Image Helpers (#606)
Added image utilities to crop an image to a 
specific region of interest plus padding as well 
as resize images to fit a max dimension. 

Identity will upload 2 images for each identity 
document side:
1. An image cropped to a bounding box 
returned by our ML model, plus some padding 
to account for ML error
2. An un-cropped image 

Both of these images will also have a maximum 
dimension that our backend expects it not to 
exceed, so the client will need to scale down the 
images to the specified dimension if it's larger.