IDPROD-2949 part 2: List View for document upload (#643)
A new `ListView` class that displays a short list
of options to the user.

This will be used for the document upload screen
in the case that the user need to select upload a
file from the device's file system (instead of
using the live camera stream to capture a
document). It will also eventually be used for
the document selection screen.