IDPROD-2949 part 3: Add timeout to document capture screen (#655)
Adds a timeout to the document capture screen as
well as options to mock the timeout until we have
the ML model and camera hooked up to the view
controller. 

The timeout period is configured on the server,
in milliseconds. When scanning a document, if the
timeout is triggered the screen will display an
error message and give the user the option to
either try again or upload an image from their
device.