IDPROD-2482 part 2: IDDetector Model output parsing (#694)
Adds logic to process an image through the
IDDetector ML model and parse its output.

The `IDDetector` class is a light-weight wrapper
that utilizes Apple's Vision framework to create
a `VNCoreMLRequest` configured specifically for
the IDDetector model. The vision observations
returned in the request's handler are parsed into
an `IDDetectorOutput` struct containing the
model's predictions and bounding box of the
detected ID document.

The current version of the IDDetector model
outputs 392 predictions of the document's
classification and bounding box. We're using the
non-maximum suppression algorithm to only
consider the best non-overlapping prediction(s).
We also filter out any predictions that return a
bounding box outside of the bounds of the images
scanned region since it means the bounding box
may not accurately reflect the actual bounds of
the ID document.