Identity example app branding customization (#721)
Adds a toggle to the Identity example app's
playground to test branding customization.

The initial release of the
IdentityVerificationSheet will try to inherit the
app's branding colors + fonts by utilizing
`tintColor` and `UIAppearance`. This change adds
a toggle in the playground that overrides the
app's tint color and UILabel appearance so that
we can easily test that our UI responds to the
app's styling. If the app does not have a custom
UILabel appearance set, we will use the default
system font.