IDPROD-3304 (bugfix): Encode 2-digit floats as strings (#794)
Stripe's API enforces that decimal numbers be
limited to 2 decimals or an API error is
returned. Sometimes (maybe 1/15 times) while
saving and ID document, the server would return
an error like:
> Invalid decimal: 0.8100000000000001

Although we were encoding `TwoDigitDecmal` to
json with the correct number of decimal places as
`0.81`, STPAPIClient uses JSONSerialization to
decode it to a Dictionary and then encode to form
data. The JSONSerialization was deserializing
this number to a Double `0.8100000000000001`
which was then being sent to the server.

This change fixes the issue by formatting the
number to a String and then encoding the String.
Since all API requests use form data, the end
result is the same, regardless of whether it's
encoded it as a string vs. a Decimal
(`.front_card_score=0.81`).