Configure AppInfo for IdentityAPIClient (#814)
Creates an `IdentityAPIClientImpl` implementation
of the `IdentityAPIClient` protocol which wraps
an `STPAPIClient` and configures the beta headers
and appInfo. This also simplifies the Identity
code by making all of the networking logic
self-contained in `IdentityAPIClient`.

Because none of the Identity API requests use a
public key but instead use an ephemeral key
secret, we're able to instantiate an APIClient
using the ephemeral key secret in place of the
public key.