IDPROD-3267 followup: Wait for non-blurry image based on time instead of frame count (#813)
This change removes the frame minimum from the
MotionBlurDetector in place of a time minimum for
consecutive frames. This is more consistent
because different devices have different frame
rates, depending on their CPU or whether they
contain a neural engine. Previous to this change,
we were still seeing some motion blur in images
on faster performing devices.

The current default is 500ms, however this will
eventually be server driven.