SelfieCaptureView (#1093)
Adds `SelfieCaptureView` which can have a state
of either displaying an error or a selfie
scanning view. This will be the content view of
the new SelfieCaptureViewController to take
selfies.

Also fixes a layout bug in `SelfieScanningView`
where the consent checkbox and label did not have
the correct horizontal insets.