Selfie Image Upload Part 3: Upload selfie images (#1159)
Crops, resizes, and uploads captured selfie
images to the server using the file upload
endpoint. A total of 6 images are uploaded to the
server consisting of a cropped high resolution
and uncropped low resolution image each for the
first, last, and best middle image of the
sequence of selfie images captured.

Also fixes a bug where the standard deviation of
the sample images was not being computed
correctly. Instead of computing the standard
deviation across all the predictions for one
image, we should compute the standard deviation
of the best prediction from each image of the
sample set. This value is not currently being
used, but it will be saved to the server in a
followup change.