Fallback to browser if VerificationSession isn't supported by client (#1164)
If a VerificationSession requires a feature that
is not supported by the mobile SDK, fallback to a
browser view. 

If the device is running iOS 14.3 or higher, the
session will be opened in a
`VerificationFlowWebViewController` which uses a
`WKWebView` to monitor the URL and detect if the
user successfully finished the session. 

Because WKWebView does not have camera access on
devices less than 14.3, the session will be
opened in an `SFSafariViewController` on lower
iOS versions. Because we cannot monitor the URL
of an SFSafariViewController, the client performs
a GET against the `verification_pages/:id`
endpoint again to check the verification
session's submission status when the user closes
the view controller.