Don't crash app if analytic can't be serialized to JSON (#1235)
An `NSInvalidArgumentException` occurs when
attempting to serialize
`JSONSerialization.data(withJSONObject:)` is used
to serialize an object that can't be serialized
to JSON, rather than throwing an error. This
change adds calls `isValidJSONObject` to ensure
the dictionary is serializable before attempting
to call `data(withJSONObject:)`.

This change also updates the analytic payload to
use a string value instead of enum, which caused
the JSON serialization to fail.