[Identity] Don't call completion block when presenting camera (#1261)
The IdentityVerificationSheet completion block
was being called early when presenting a
UIImagePickerController when manually uploading a
document photo because it was getting triggered
on the viewDidDisappear callback on the
navigation controller. This fixes the issue by
only calling the completion block when the
navigation controller is not currently presenting
another view controller.