Add iOS Jira component to notify_ci script (#2374)
Right now, CI failures aren't tagged with a component in our Jira bucket. This makes it difficult to split out run metrics by component (iOS vs. Android, for example). While we could add the component when we resolve (I just did this for all previously created tickets as a bulk operation in Jira), having a robot do it for us is much nicer.