Add validation to catch blank ephemeral keys in CustomerConfiguration
We currently don't validate that an ephemeral key is non-empty when a
user includes it in the `CustomerConfiguration`. This ends up giving
users a pretty obscure error in the API response, and is difficult to
debug today.

Including this validation will catch this issue early and provide users
a better debugging experience.