Enable testmode for Stripe Dashboard app (#2849)
Allows the Stripe Dashboard app to internally test the payment sheet in test mode.

The Stripe Dashboard app has an internal debug menu that allows us to toggle live mode vs. test mode during development, however the SDK was setting the `Stripe-Livemode` header needed for Dashboard keys based on a run scheme environment variable instead of the debug menu. The env var could get out of sync if the debug menu was changed and would only work when running the app attached to Xcode.

This fix enables the Dashboard app to directly toggle livemode in the SDK when importing the module with `@_spi(DashboardOnly)`.