Fix PayNow animation jank (#2934)
## Summary
When the PaymentSheetViewController appears, it refreshes its UI,
including the ConfirmButton's animation state. A race condition arises
when the PaymentSheetViewController and the completion of an action
simultaneously update the UI, resulting in occasional animation
glitches. These occur when we start new animations before the
PaymentSheet animations have completed. To resolve this issue, we delay
the action's evolution to succeeded/cancelled until 0.2 seconds after
the dismissal of the PollingViewController. This allows sufficient time
for the animations to finish, preventing conflicts.

## Motivation
- Polish

## Testing
- Manually ran it 25 times and no longer observed the animation jank

## Changelog
N/A