Fix PaymentSheet picker on macOS Catalyst (#2948)
## Summary
Pickers in PaymentSheet (including the country selector) don't work, as
UITextField inputViews don't appear in Catalyst. This means clicking the
picker will do nothing.

I evaluated a few different alternatives for this, and the least-bad one
seems to be a UIButton with an attached menu. Ideally we'd have some
sort of NSComboBox equivalent on Catalyst, but I haven't been able to
find a publicly available one yet.

UIMenu is only available on macOS 11, but Catalyst wasn't very usable on
10.15 (our minimum version in the README). Looking at our analytics, we
don't appear to have any users before macOS 12, so I feel comfortable
using this and bumping the version in the README.

## Motivation
Fixing country selection on Catalyst.

## Testing
In PaymentSheet Example

## Changelog
Added