Fix card element not updating value when deleting on empty field (#2982)
## Summary
This is a followup to https://github.com/stripe/stripe-ios/pull/2971, it
turns out that on top of not calling the delegate, we were not updating
the backing view model, this was due to the implementation of
`UITextField.deleteBackwards()`, which doesn't update the `text`
property directly, causing `didSet` to not be called, which we depend on
to update the backing view model. To get around this, I'm manually
removing the last character in the text if it isn't empty.

## Motivation
https://github.com/stripe/stripe-ios/issues/2483

## Testing
- Verified visually
- Added test

## Changelog
* [Fixed] An issue with `STPPaymentCardTextField`, where the card params
were not updated after deleting an empty sub field.