Pass email to Link URL generator (#3071)
## Summary
For the LinkPaymentController bank-only flow, we want to pass the
billing details in the email address to the Link URL generator.

## Motivation
Autofilling emails in the deferred intent flow.

## Testing
LinkPaymentController doesn't appear to have any tests. Will follow up
with the product owners later.

## Changelog
None, this is a private beta feature.