Unblock building on visionOS (#3078)
## Summary
Committing the bare minimum changes to get StripePaymentSheet building
on visionOS.
* Makes the needed project changes for visionOS
* Conditionally removes most uses of UIScreen
* Disables iOSSnapshotTestCase when building test targets on visionOS,
as it isn't supported. Created a no-op version of STPSnapshotTestCase so
that we don't have to `ifdef` every snapshot test.
* Switches to iOS 16+ Locale APIs when running on visionOS
* Removes any code related to trait changes. We can implement these
later, but as far as I can tell visionOS doesn't change traits on live
views. (yet!) The new trait observer API is a lot of extra code, so we
can probably hold off until this is needed.
* Removed UIButton insets (we may need to investigate these later,
there's a new API for them)
* Disabled STPCardScanner (there's no camera!)
* Conditionally compiled out any UIFeedbackGenerators
* Conditionally compiled out SFSafariViewControllerDelegate (Safari runs
in a separate app, like in macCatalyst)

## Motivation
visionOS support

## Testing
Added a Bitrise test to ensure it continues to compile, but I don't
expect it will fully work yet. Will continue adding tests.

## Changelog
Not yet.