Fix STPApplePayContext initializer not fully validating the payment r… (#3097)
…equest

## Summary
STPApplePayContext's init says it returns nil if the request is invalid.

It does this, in part, by checking if `PKPaymentAuthorizationController`
is nil.

PKPaymentAuthorizationController's docs state:
"If the user can’t make payments on any of the payment request’s
supported networks, initialization fails and this method returns nil."

In actuality, this initializer is non-nullable. To make sure
STPApplePayContext returns nil when the request is invalid, we'll use
PKPaymentAuthorizationViewController's initializer, which *is* nullable.


## Motivation
https://github.com/stripe/stripe-ios/issues/2937

## Testing
See unit test - failed before, passes after.

## Changelog
### Apple Pay
* [Fixed] STPApplePayContext initializer returns nil in more cases where
the request is invalid.