Support visionOS across Xcode 14 and 15 (#3127)
## Summary
We need to support visionOS across Xcode 14 and 15 and all package
managers. To do this, we need a solution that doesn't involve
compile-time flags, as SPM won't let us provide them.

See
https://github.com/stripe/stripe-ios/pull/3127/files#diff-ae2869cf29a9ac2a97e2b2f976d80f6ec5b7e59a004b6bf6bee0aa2e3f281ba5R148
for a full explanation of the hack.

## Motivation
Fix visionOS support in various package managers.

## Testing
CI and manually in Xcode 15.3. Re-adding automated visionOS tests in
https://github.com/stripe/stripe-ios/pull/3126.

## Changelog
Added.