Link Passthrough V2 (#3159)
## Summary
* Pipe the `link_passthrough_mode_enabled` flag to the Link popup.
* Remove some dead test code.
* Make LinkEnabledPaymentMethodElement a ContainerElement, fixing some
code in PaymentSheet that relies on the ability to traverse elements.
(e.g. filling the address fields based on an AddressViewController)
* Adds support for signing up for Link in passthrough mode via the
native flow. In this situation, we `/share` the payment details as a
card payment method, then confirm the PI as usual with that PM ID.

## Testing
Manually in PaymentSheet Example, bringing up end-to-end signup test.

Note that passthrough mode will not be offered on the current SDK
version because of a server-side gate.

## Changelog
None.