Remove use of keychain APIs (#3191)
## Summary
We've removed all sensitive settings from Link, so it doesn't make sense
to use the keychain anymore. Remove it and use UserDefaults for the one
remaining bool.

## Motivation
Avoids keychain access errors when using multiple apps from the same
developer. I was originally planning to fix this by appending an App ID
to the keys, but at this point it makes more sense to delete this unused
code instead.

## Testing
CI & manual Link transaction in PaymentSheet Example