Financial Connections: introduced a new ErrorViewController for unexpected_error pane and added it to PartnerAuth related errors (#3277)
## Summary

This PR:
- Creates a new "pane" (/screen) called `ErrorViewController` driven by
the pane `unexpected_error`
- It then uses this new "pane" to display errors that used to come from
`PartnerAuthViewController` (the reason for this code change is because
we moved logic from `PartnerAuthViewController` to
`InstitutionPickerViewController` (for sheets)

---

NOTE: This PR is merged into a new feature branch (`kg-v3nav`) as there
will be more navigation-related changes similar to this. But will be
merged into the `fc-v3` feature branch after `kg-v3nav` is ready.

Details to understand before reviewing PR:
- This is one PR, out of expected ~50+, to change the design of
Financial Connections from a "V2" design to a "V3" design. [Here is a
link to the
designs](https://www.figma.com/file/wXQ8NJApqSJiLmxxANDRTs/%E2%9C%A8-Connections-3.0?type=design&node-id=2524-217093&mode=design&t=kXaBiAlgVsC7YcH1-0).
- The code here is not the finished version. It's just an iteration
towards the final version. This code will be re-visited later. The goal
is to go screen-by-screen and make easy changes to learn about all the
difficulties in each screen. Later, all the learnings will be combined
to do a "final" polish of all screens.
- The above also applies to design. The design here is not the finished
version.
- This PR is merged into a feature branch `fc-v3`. It is not merged into
`master`.
- There might be TODO's scattered throughout the code. That's expected
and they will be fixed later.

## Testing

This video shows errors now appearing from the `createAuthSession` call
of `InstitutionPickerViewController` (which is also reused in
`PartnerAuthViewController`)


https://github.com/stripe/stripe-ios/assets/105514761/d40bc30a-cec1-4d44-97d6-246719e2146d