Financial Connections: for v3, worked on supporting app to app state for sheets (#3285)
## Summary

This PR:
- Adds "app to app" support for the new sheet V3 design. "App to app" is
the ability to connect a bank account directly through the banking app
(ex. Chase) rather than an OAuth Website.

---

NOTE: This PR is merged into a new feature branch (`kg-v3nav`) as there
will be more navigation-related changes similar to this. But will be
merged into the `fc-v3` feature branch after `kg-v3nav` is ready.

Details to understand before reviewing PR:
- This is one PR, out of expected ~50+, to change the design of
Financial Connections from a "V2" design to a "V3" design. [Here is a
link to the
designs](https://www.figma.com/file/wXQ8NJApqSJiLmxxANDRTs/%E2%9C%A8-Connections-3.0?type=design&node-id=2524-217093&mode=design&t=kXaBiAlgVsC7YcH1-0).
- The code here is not the finished version. It's just an iteration
towards the final version. This code will be re-visited later. The goal
is to go screen-by-screen and make easy changes to learn about all the
difficulties in each screen. Later, all the learnings will be combined
to do a "final" polish of all screens.
- The above also applies to design. The design here is not the finished
version.
- This PR is merged into a feature branch `fc-v3`. It is not merged into
`master`.
- There might be TODO's scattered throughout the code. That's expected
and they will be fixed later.
## Testing

### Test Native App To App

This video shows that the "app to app" works for Native (with the UI
state changing):


https://github.com/stripe/stripe-ios/assets/105514761/7295a36c-16d0-4a81-ade5-4f664fbede0c

### Test Web App To App

This video shows that the "app to app" works for Web (with the UI state
changing):


https://github.com/stripe/stripe-ios/assets/105514761/e6f83003-ec5c-4105-8ccf-7a65b33a4cac