Fix test playground + customerID settings (#3343)
# Summary
Test playground was reloading too many times in tests unnecessarily.
This is because load was modifying settings.customerId, which was
causing it to reload (when autoreload is on).

The problem is that load completion should never touch settings, that's
an infinite loop waiting to happen. It doesn't make sense for customerID
to be in the settings object anyways, since it's not a setting - it
can't be set, it's just a readonly view of the current customer ID.


# Motivation
Was looking at why UPI tests were flaking sometimes
https://jira.corp.stripe.com/browse/RUN_MOBILESDK-2975

I _think_ there's a possible cause that looks like:

We modify settings.customerID when the load finishes -> we stop spinning
and show the "Present PaymentSheet" -> The test sees "Present
PaymentSheet" become enabled -> we autoreload because the settings
changed -> we start spinning -> The test tries to tap "Present
PaymentSheet" and fails

Hopefully this PR reduces/eliminates the flake.