Adjust layout for consumer assets (#3384)
## Summary
* Added latest consumer assets behind a flag. The flag intentionally
defaults to `true`, but it currently set to `false` server-side.
* Updated snapshot tests to use new assets, added legacy tests for older
assets
* Added a singleton (I'm sorry) to set this state from the
`elements/sessions` endpoint. This made more sense than trying to pass
the bool down to each view in the hierarchy and the public-facing
PaymentOption. We'll get rid of the singleton once the flag is removed
server-side in the next few months.
* Removed some unused Link UI theme, and stopped using the Link UI theme
for the OTP UI snapshot test (as it was only used in that test)

## Testing
Existing snapshot tests, CI