Refactor `DownloadManager` and add analytics (#3487)
## Summary
### DownloadManager refactor
- Removed conformance to `URLSessionDelegate`, it was never implemented
or used.
- Removed various variables relating to storing and keeping track of
pending requests. Usage of Swift concurrency takes care of this since
the functions "look" like they are executing synchronously. Some of
these such as the `downloadQueue` and `downloadOperationQueue` were
never used.
- Removed `STPCacheExpirationInterval`, it was never used. Instead the
`URLCache` uses the
[`Cache-Control`](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cache-Control)
HTTP header to determine when an item will expire.
- Created a common function `downloadImage` which uses Swift async to
download images and catch errors in a single location
- Uses `data` instead of `downloadTask` on the `URLSession` as this is a
small operation that does not need to occur in the background, and
allows us to not worry about converting the response to data which could
result in an additional error we need to handle.
- Refactored the in-memory cache to not use GCD and instead use Actor
pattern.
- The public API of `DownloadManager` is not modified/refactored to
avoid refactoring all usage of the class, kept the refactor internal. We
should consider removing the blocking image download and update the
usages.
- URL Cache is thread safe:
https://developer.apple.com/documentation/foundation/urlcache#2317483

### Analytics changes
- If we get an error back from the network we log that error so that
`error_type` = `NSURLErrorDomain` and `error_code` is the error code
from the response.
- If we fail to convert the data into an image we send `error_type` =
`StripeCore.DownloadManager.Error` and `error_code` =
`failedToMakeImageFromData`

## Motivation
Feedback

## Testing
- Manual in PaymentSheet playground
- Existing tests ran 1k times each
- Few new tests

## Changelog
N/A