Update PaymentIntent+API.swift
I made this change because, without it, it is not possible to complete the payment intent in Stripe. The issue occurred during the transaction at the endpoint /paymentintent/{id}/confirm, as the Stripe API did not accept the client_secret in the same request. With this modification, the library now functions correctly, allowing for the successful confirmation of the payment intent.