Update .`xctestplan`s to use US region and locale (#3675)
## Summary

Since I'm based outside of the US, a lot of the UI components are
localized slightly differently for me. This caused a lot of UI
integration tests to fail.

In order to fix this, and to provide a consistent experience for
everyone writing UI tests, this updates all the `.xctestplan` files to
specify a US locale.

For example, a lot of integrations tests in the Payments Sheet module
look for the charge button and tap it, which looks like this:

```swift
app.buttons["Pay $50.99"].tap()
```

However, here in Canada, the label actually says `Pay US$50.99`, so this
step was always failing.

<img width="396" alt="Screenshot 2024-06-17 at 2 32 03 PM"
src="https://github.com/stripe/stripe-ios/assets/172562065/eb69accc-e4dd-4992-a486-69175ad3b002">

Another example is in the billing address form, instead of the `ZIP
code` field, we show `Postal code`.

In the second commit, I fix a couple flaky integration tests.

## Testing

There should be no functional changes, and all unit / UI / integration
tests should continue to pass.