[FC] Theme-based header logo (#3772)
## Summary

This adds support for theme-based header logos in the Financial
Connections flow. Right now we support Link for the `linkLight` theme,
and continue to use the Stripe logo for all other themes.

## Motivation

Ultimately, the goal here is to show the Link logo in the Instant Debits
flow, but is also extensible to support any other themes that might
come.

## Testing

Manually verified that the link logo is only shown for the instant
debits flow, and that the FC flow still shows the stripe logo. Also
verified this still works with the test mode badge:


https://github.com/user-attachments/assets/f64307c8-fa14-4438-acd4-23ef0915a410

## Changelog

N/a