[BANKCON-11512] Add killswitch for Native Instant Debits flow (#3823)
Resolves https://jira.corp.stripe.com/browse/BANKCON-11512

## Summary

This makes use of the native experience killswitch
(`bank_connections_mobile_native_version_killswitch` in the manifest's
`feature` payload) in the native instant debits flow router. If the
killswitch is active, we will always show the web instant debits flow.

In the second commit, I also added some unit tests for the `FlowRouter`
to make sure everything works as expected.

## Motivation

The killswitch will allow us to remotely deactivate the native
experience if we detect an issue with it.

## Testing

Unit tests added! 

Also in this video you can see that with the `Automatic` flow selected,
the Networking merchant, which does not have the killswitch activated
(you can see the test merchant `carlosmuvi-us`
[here](https://amp.corp.stripe.com/feature-flags/flag/bank_connections_mobile_native_instant_debits_eligible_merchants)),
but for a different merchant which has the killswitch activated, they
get the web flow:


https://github.com/user-attachments/assets/5c170605-f917-4bc1-92de-7c47af3a4910


## Changelog

N/a