[BANKCON-11811] Add handler for link-account-picker URL on Consent pane (#3904)
## Summary

As part of the Networking Manual Entry project, this adds handling for
`stripe://link-account-picker` URLs in the server-driven text shown on
the Consent pane. This is specifically for the scenario when a user has
signed in to Link on a hosted surface and then launches the auth flow.

## Motivation

Reflects
https://git.corp.stripe.com/stripe-internal/stripe-js-v3/pull/23737

## Testing

Hard to test without being able to reproduce the scenario where a user
is signed in to Link on a hosted surface yet. Though I can verify that
other Stripe-scheme link handling did not break with this change:


https://github.com/user-attachments/assets/d8d00c5a-7a78-4c72-9780-bb1c8d5be90e

## Changelog

N/a